Add files via upload
We created a tool that uses a data-driven approach to provide authors with insights into the readability of their published scholarly work with regard to other text repositories. The tool uses an entered author's name to perform a web scrape of their scientific work. It then quantifies an existing text repository with complexity shown to be comparable to that of other scientific journals. The tool subsequently uses this output as a reference to show how the readability of user-selected written work compares to this source.

Ultimately, this tool will expand upon current readability metrics by computing a more detailed and comparative look at the complexity of written text. We hope that this will allow scientists and other experts to better monitor the complexity of their writing relative to other text types, leading to the creation of more accessible online material. And perhaps more broadly contribute to an improved global communication and understanding of complex topics.